+++ openerp/service/report.py
@@ -5,8 +5,8 @@ import logging
 import sys
 import threading
 
-import openerp.netsvc
-import openerp.pooler
+import openerp
+import openerp.report
 from openerp import tools
 
 import security
@@ -49,10 +49,9 @@ def exp_render_report(db, uid, object, ids, datas=None, context=None):
 
     self_reports[id] = {'uid': uid, 'result': False, 'state': False, 'exception': None}
 
-    cr = openerp.pooler.get_db(db).cursor()
+    cr = openerp.registry(db).db.cursor()
     try:
-        obj = openerp.netsvc.LocalService('report.'+object)
-        (result, format) = obj.create(cr, uid, ids, datas, context)
+        result, format = openerp.report.render_report(cr, uid, ids, object, datas, context)
         if not result:
             tb = sys.exc_info()
             self_reports[id]['exception'] = openerp.exceptions.DeferredException('RML is not available at specified location or not enough data to print!', tb)
@@ -88,10 +87,9 @@ def exp_report(db, uid, object, ids, datas=None, context=None):
     self_reports[id] = {'uid': uid, 'result': False, 'state': False, 'exception': None}
 
     def go(id, uid, ids, datas, context):
-        cr = openerp.pooler.get_db(db).cursor()
+        cr = openerp.registry(db).db.cursor()
         try:
-            obj = openerp.netsvc.LocalService('report.'+object)
-            (result, format) = obj.create(cr, uid, ids, datas, context)
+            result, format = openerp.report.render_report(cr, uid, ids, object, datas, context)
             if not result:
                 tb = sys.exc_info()
                 self_reports[id]['exception'] = openerp.exceptions.DeferredException('RML is not available at specified location or not enough data to print!', tb)